converter: move out encryption package
Due to the introduction of the github.com/containers/ocicrypt package
in converter, which brings in many other dependencies, but encryption
should currently be an experimental optional feature, we need to move
this part of the codes out to reduce the burden on other projects that
reference the package.
