+++ pkg/encryption/encryption.go
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package converter
+package encryption
 
 import (
 	"context"
